Move shortcuts and locale manager to services folder [ci skip]
