+++ src/lib/services/locale-manager.svelte.ts
