+++ src/lib/services/shortcuts.svelte.test.ts
