+++ src/lib/services/shortcuts.svelte.ts
